updated ctrl
