+++ models/contacts.json
@@ -64,5 +64,11 @@
     "name": "Anna Red",
     "email": "anna@gmail.com",
     "phone": "(044) 555-8673"
+  },
+  {
+    "id": "T8fO1rAuvzuyPhoAkgwOY",
+    "name": "Anna Novak",
+    "email": "anna@gmail.com",
+    "phone": "(044) 555-8673"
   }
 ]
